moved examples for 'removing all occurrences of an element from a list' to core-java-collections
